+++ src/controllers/index.js
@@ -1,11 +1,9 @@
 const userController = require("./userController");
 const threadController = require("./threadController");
-const threads2Controller = require("./threads2Controller");
 const likeController = require("./likeController");
 
 module.exports = {
   userController,
   threadController,
-  threads2Controller,
-  likeController,
+  likeController
 };
