+++ src/models/index.js
@@ -1,11 +1,9 @@
 const threadDao = require("./threadDao");
-const threads2Dao = require("./threads2Dao");
 const userDao = require("./userDao");
 const likeDao = require("./likeDao");
 
 module.exports = {
   threadDao,
-  threads2Dao,
   userDao,
-  likeDao,
+  likeDao
 };
