+++ src/services/index.js
@@ -1,11 +1,9 @@
 const threadServices = require("./threadServices");
-const threads2Services = require("./threads2Service");
 const userService = require("./userService");
 const likeService = require("./likeService");
 
 module.exports = {
   threadServices,
-  threads2Services,
   userService,
-  likeService,
+  likeService
 };
