+++ src/services/threadServices.js
@@ -12,9 +12,37 @@ const threadWrite = async (newThreadUser, newThreadContent) => {
   await threadDao.insertThread(newThreadUser, newThreadContent);
 };
 
+const threadsList = async () => {
+  const threadsListData = await threadDao.threadsList();
+
+  return threadsListData;
+};
+
+const modifyThreads = async (foundUser, threadId, content) => {
+  const thread = await threadDao.findThreadById(threadId);
+
+  //TypeError: Cannot read property 'user_id' of undefined도 처리하기 위해 || typeof thread[0] === 'undefined' 추가
+  if (!thread || typeof thread[0] === "undefined") {
+    const error = new Error("THREAD_NOT_FOUND");
+    error.status = 404;
+    throw error;
+  }
+
+  // 작성자 = 로그인한 사용자인지
+  if (thread[0].user_id !== foundUser.id) {
+    const error = new Error("FORBIDDEN_USER");
+    error.status = 403; // 403 : Forbidden 클라이언트가 요청한 리소스에 접근 권한 X
+    throw error;
+  }
+
+  await threadDao.modifyThreads(threadId, content);
+  // 에러가 발생하지 않았다면 클라이언트에게 성공 응답
+  return { success: true, message: "update success" };
+};
+
 const threadDelete = async (threadId, userId) => {
   //  Error handling - 존재하지 않는 쓰레드를 삭제하는 경우
-  const existingThread = await userDao.findThread(threadId);
+  const existingThread = await threadDao.findThreadById(threadId);
   console.log(existingThread);
   if (existingThread.length == 0) {
     const err = new Error("CONTENT_NOT_FOUND");
@@ -24,7 +52,7 @@ const threadDelete = async (threadId, userId) => {
   }
 
   //  Error handling - 다른 유저가 쓴 쓰레드를 삭제하는 경우
-  const thisUserThread = await userDao.findThreadUser(userId);
+  const thisUserThread = await threadDao.findThreadUser(userId);
   console.log(thisUserThread);
   if (existingThread.length > 0 && thisUserThread.length == 0) {
     const err = new Error("NOT_YOUR_THREAD");
@@ -36,4 +64,4 @@ const threadDelete = async (threadId, userId) => {
   await threadDao.deleteThread(threadId, userId);
 };
 
-module.exports = { threadWrite, threadDelete };
+module.exports = { threadWrite, threadsList, modifyThreads, threadDelete };
